backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
python-multipart
python-dotenv
alembic
pydantic
pydantic-settings
